Adding transaction field to some *Options interfaces
By looking at the implementation, all options end up being passed to the general query function, which accepts options with transactions.

To guarantee that a transaction is respected, all methods of query interface must have the same transaction on options. However, the types given exclude this field for some reason.

This commit is made exclusively to fix the use case at hand (add index/constraint). If needed, it may be extended to the whole file.